requirements.txt
langchain
langchain-openai
openai
python-dotenv
streamlit
pypdf2
python-docx

